📝 docs(README.md): add authenticate method documentation to provide clear instructions on how to authenticate with the Colissimo API 🔧 refactor(index.ts): extract types and constants to separate files for better code organization and readability ✨ feat(index.ts): add authenticate method to ColissimoAPI class to support authentication with the Colissimo API 🚚 move(ENDPOINT): move ENDPOINT constant to constants.ts file for better code organization 📦 new(types.ts): add new types file to hold all type definitions for better code organization and readability
🔧 refactor(index.test.ts): restructure imports for better readability
✨ feat(index.test.ts): add authenticate test cases to increase test coverage
🔨 fix(index.test.ts): correct axios post url in findRDVPointRetraitAcheminement test cases for accurate testing
