+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const ENDPOINT = 'https://ws.colissimo.fr'
